refactor(auth): unify login process into a single page
Replaces the two-step email/password login flow with a single LoginPage component. Updates routing accordingly.
