+++ fronted/src/style/UserLogin.css
@@ -2,80 +2,99 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  height: 100vh;
+  min-height: 100vh;
+  padding: 1rem;
 }
 
 .login-container {
-  background: #37437C;
-  padding: 2rem 2.5rem;
-  border-radius: 1rem;
-  box-shadow: 0 10px 25px rgba(0, 0, 0, 0.35);
+  background: rgba(255, 255, 255, 0.1);
+  backdrop-filter: blur(15px);
+  padding: var(--spacing-xl) var(--spacing-2xl);
+  border-radius: var(--radius-2xl);
+  border: 1px solid rgba(255, 255, 255, 0.2);
+  box-shadow: var(--shadow-xl);
   width: 100%;
-  max-width: 380px;
+  max-width: 420px;
+  color: var(--text-white);
 }
 
 .login-container h1 {
   text-align: center;
-  margin-bottom: 1.5rem;
-  color: #fff;
+  margin-bottom: var(--spacing-lg);
+  color: var(--text-white);
+  font-family: var(--font-archivo);
 }
 
 .login-container input {
   width: 100%;
-  padding: 0.75rem 1rem;
-  margin-bottom: 1rem;
-  border: none;
-  border-radius: 0.5rem;
-  background: #46569E;
-  color: #fff;
-  font-size: 1rem;
-  transition: 0.3s ease;
+  padding: var(--spacing-md) var(--spacing-lg);
+  margin-bottom: var(--spacing-lg);
+  border: 1px solid rgba(255, 255, 255, 0.3);
+  border-radius: var(--radius-md);
+  background: rgba(0, 0, 0, 0.2);
+  color: var(--text-white);
+  font-size: var(--text-base);
+  transition: all var(--transition-base);
+  font-family: var(--font-sans);
 }
 
 .login-container input::placeholder {
-  color: #cfd3ff;
+  color: rgba(255, 255, 255, 0.6);
 }
 
 .login-container input:focus {
   outline: none;
-  background: #5568C0;
-  box-shadow: 0 0 0 2px #677FEB;
+  background: rgba(0, 0, 0, 0.3);
+  border-color: var(--color-primary-1);
+  box-shadow: 0 0 0 3px rgba(99, 123, 226, 0.3);
 }
 
 .login-container button {
   width: 100%;
-  padding: 0.85rem 1rem;
+  padding: var(--spacing-md) var(--spacing-lg);
   border: none;
-  border-radius: 0.5rem;
-  background: #677FEB;
-  color: #fff;
-  font-size: 1rem;
+  border-radius: var(--radius-md);
+  background: var(--color-primary-1);
+  color: var(--text-white);
+  font-size: var(--text-base);
   font-weight: bold;
   cursor: pointer;
-  transition: 0.3s ease;
+  transition: all var(--transition-base);
+  font-family: var(--font-archivo);
+  text-transform: uppercase;
 }
 
 .login-container button:hover {
-  background: #5568C0;
+  background: var(--color-primary-2);
+  transform: translateY(-2px);
+  box-shadow: var(--shadow-lg);
 }
 
 .login-container button:active {
-  background: #46569E;
+  background: var(--color-primary-3);
+  transform: translateY(0);
 }
 
-.email-container .create-account {
+.create-account {
   text-align: center;
-  margin-top: 1rem;
+  margin-top: var(--spacing-lg);
 }
 
-.email-container .create-account a {
-  color: #cfd3ff;
+.create-account p {
+  color: rgba(255, 255, 255, 0.8);
+  margin-bottom: var(--spacing-xs);
+  font-size: var(--text-sm);
+}
+
+.create-account a {
+  color: var(--text-white);
   text-decoration: none;
-  font-size: 0.9rem;
+  font-size: var(--text-sm);
   font-weight: bold;
-  transition: 0.3s ease;
+  transition: color var(--transition-base);
 }
 
-.email-container .create-account a:hover {
-  color: #fff;
+.create-account a:hover {
+  color: var(--color-primary-1);
+  text-decoration: underline;
 }
